dbc: widen name_off to uint32_t, sig_count to uint16_t for large DBC support
The Tesla Model 3/Y dedup DBC (403 messages, 8881 signals) pushes
the string table past 64 KiB, overflowing uint16_t name_off.
Several messages also exceed 255 signals, requiring uint8_t -> uint16_t.

Breaking binary format change — DBC_VERSION bumped to 2.
- dbc_types.h: name_off uint32_t, sig_count uint16_t, version 2
- dbc_pack.h: dbc_str() parameter uint32_t
- dbc2bin.py: updated struct.pack formats for version 2
- vehicle_dbc_embedded.h: regenerated from tesla_model3_vehicle.dbc
- tests: offset arrays widened to uint32_t

+++ firmware/src/dbc/dbc_pack.h
@@ -39,7 +39,7 @@ int dbc_find_signal_by_msg_name(const dbc_t *dbc, const char *msg_name, const ch
 /* Look up a message by its name. Linear scan; fine for typical DBC sizes. */
 const dbc_msg_t *dbc_find_msg_by_name(const dbc_t *dbc, const char *name);
 
-static inline const char *dbc_str(const dbc_t *dbc, uint16_t off)
+static inline const char *dbc_str(const dbc_t *dbc, uint32_t off)
 {
     return dbc->strs + off;
 }
